用户可以删除自己发的话题

+++ app/Http/Controllers/TopicsController.php
@@ -103,9 +103,10 @@ class TopicsController extends Controller
      */
     public function destroy(Topic $topic)
     {
+        $this->authorize('destroy', $topic);
         $topic->delete();
 
-        return redirect()->route('topics.index');
+        return redirect()->route('topics.index')->with('success', '删除成功！');
     }
 
     /**
